+++ tests/test_set.rs
@@ -8,21 +8,14 @@ use rstest_reuse::{self, *};
 
 /// Test all shell integrations with a simple `es set` command
 #[apply(all_shells)]
-fn test_subcommand_set(
+fn test_set_subcommand(
     shell_kind: &str,
     #[values(false, true)] detect_shell: bool,
 ) {
     execute_script(
         "
-        es set integration-tests p1
-
-        echo -n \
-            $VARIABLE1 \
-            $VARIABLE2 \
-            $VARIABLE3 \
-            $VARIABLE4 \
-            $FILE_VARIABLE1 \
-            $FILE_VARIABLE2
+        es set test p1
+        echo -n $VAR1 $VAR2 $VAR3 $FILE_VAR1 $FILE_VAR2
         ",
         shell_kind,
         detect_shell,
@@ -35,11 +28,10 @@ pre setup 2
 post setup 1 abc
 post setup 2 abc
 The following variables will be set:
-VARIABLE1 = abc
-VARIABLE2 = def
-VARIABLE3 = ghi
-FILE_VARIABLE1 = 123
-abc def ghi 123",
+VAR1 = abc
+VAR2 = def
+FILE_VAR1 = 123
+abc def 123",
     )
     .stderr("");
 }